requirements.txt
#Frameworks

Django==4.2.13
djangorestframework==3.15.1

#Authentication & Security

djangorestframework-simplejwt==5.3.1

#Deployment

gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9  # (for Postgres if deployed on Heroku)

#Environment Variables

python-dotenv==1.0.1

#Optional (CORS, testing)

django-cors-headers==4.3.1
pytest==8.2.0
pytest-django==4.8.0

django-filter>=24.2